Improved performance and compatibility

+++ scripts/killslow.sh
@@ -59,19 +59,20 @@ ban_user() {
 		[ -z "$5" ] && return 0	
 		[ -z "$6" ] && return 0		
 		[ $BANUSER -lt 1 ] && return 0			
-		cat $4/$1 | grep -P "^FLAGS " | grep 6 > /dev/null && return 0	
-		[ -n "$LOG" ] && echo "[$(date "+%T %D")] DISABLE USER: $1 for $BANUSER seconds.." >> $LOG
-		cat $4/$1 | sed -r 's/^FLAGS .*$/&6/' > /tmp/ks.$1.$$.dtm &&	
+		egrep -q '^FLAGS .*6' $4/$1 && return 0
+		[ -n "$LOG" ] && echo "[`date "+%T %D"`] DISABLE USER: $1 for $BANUSER seconds.." >> $LOG
+		sed -r 's/^FLAGS .*$/&6/' "$4/$1" > /tmp/ks.$1.$$.dtm &&	
 			cat /tmp/ks.$1.$$.dtm > $4/$1 &&			
 			$5 --sleep $BANUSER --fork "$6 unban $1 0 $4"
 		rm /tmp/ks.$1.$$.dtm
 		
 		return 0
 	elif [ $2 -eq 1 ]; then
-		! cat $4/$1 | grep -P "^FLAGS " | grep 6 > /dev/null && return 0
-		[ -n "$LOG" ] && echo "[$(date "+%T %D")] ENABLE USER: $1" >> $LOG
-		g_FLAGS=$(cat $4/$1 | grep -P "^FLAGS " | tr -d '6') && 
-			cat $4/$1 | sed  "s/^FLAGS .*$/$g_FLAGS/" > /tmp/ks.$1.$$.dtm &&
+		egrep -q '^FLAGS .*6' $4/$1 || return 0
+		[ -n "$LOG" ] && echo "[`date "+%T %D"`] ENABLE USER: $1" >> $LOG
+		g_FLAGS=`cat "$4/$1" | egrep '^FLAGS ' | tr -d '6'`
+		[ -z "$g_FLAGS" ] && return 1
+		sed -r "s/^FLAGS .*/$g_FLAGS/" "$4/$1" > /tmp/ks.$1.$$.dtm &&
 			cat /tmp/ks.$1.$$.dtm > $4/$1 &&
 			rm /tmp/ks.$1.$$.dtm	
 		
@@ -84,13 +85,13 @@ elif [[ "$1" == "unban" ]];then
 	ban_user $2 1 $3 $4 && exit 1
 fi
 
-! echo $6 | grep -P "^STOR" > /dev/null && exit 1
+echo $6 | egrep -q '^STOR' || exit 1
 
-! echo $6 | grep -P $FILES_ENFORCED > /dev/null && exit 1
-echo $9 | grep -P $PATHS_FILTERED > /dev/null && exit 1
+echo $6 | egrep -q $FILES_ENFORCED || exit 1
+echo $9 | egrep -q $PATHS_FILTERED && exit 1
 
-[ -n "$EXEMPTUSERS" ] && echo $3 | grep -P "^($EXEMPTUSERS)\$" > /dev/null && exit 1
-[ $EXEMPTSITEOPS -eq 1 ] && echo $8 | grep 1 > /dev/null && exit 1
+[ -n "$EXEMPTUSERS" ] && echo "$3" | egrep -q "^($EXEMPTUSERS)\$" && exit 1
+[ $EXEMPTSITEOPS -eq 1 ] && echo "$8" | grep -q 1 && exit 1
 
 ! [ -d "/tmp/du-ks" ] && mkdir -p /tmp/du-ks
 
@@ -104,9 +105,9 @@ fi
 DRATE=$5
 LUPDT=$2
 GLUSER=$3
-CT=$(date +%s)
+CT=`date +%s`
 
-DIFFT=$(expr $CT - $LUPDT);
+DIFFT=`expr $CT - $LUPDT`;
 
 #[ $DIFFT -lt 1 ] && exit 1
 
@@ -118,38 +119,41 @@ KILLED=0
 
 
 [ $DIFFT -gt $WAIT ] && [ $DRATE -lt $MINRATE ] && SLOW=1 && [ $VERBOSE -gt 0 ] && 
-        O="[$(date "+%T %D")] WARNING: Too slow (running $DIFFT secs): $GLUSER [PID: $4] [Rate: $DRATE/$MINRATE B/s]\n"  
+        O="[`date "+%T %D"`] WARNING: Too slow (running $DIFFT secs): $GLUSER [PID: $4] [Rate: $DRATE/$MINRATE B/s]\n"  
 
-if [ $SLOW -eq 1 ] && [ -f /tmp/du-ks/$4 ]; then
-	MT1=$(stat -c %Y /tmp/du-ks/$4) 
-	UNDERTIME=$(expr $CT - $MT1)
-	[ $UNDERTIME -gt $MAXSLOWTIME ] && 
-		O="[$(date "+%T %D")] KILLING: [PID: $4]: Below speed limit for too long ($UNDERTIME secs): $GLUSER [Rate: $DRATE/$MINRATE B/s]\n" &&  
-		SHOULDKILL=1 && ban_user $GLUSER 0 $8 ${10} $7 $0  && kill $4 && KILLED=1 &&  rm /tmp/du-ks/$4
+if [ $SLOW -eq 1 ] && [ -f "/tmp/du-ks/$4" ]; then
+	MT1=`stat -c %Y "/tmp/du-ks/$4"` 
+	UNDERTIME=`expr $CT - $MT1`
+	[ $UNDERTIME -gt $MAXSLOWTIME ] && {		
+		echo "[`date "+%T %D"`] KILLING: [PID: $4]: Below speed limit for too long ($UNDERTIME secs): $GLUSER [Rate: $DRATE/$MINRATE B/s]" >> $LOG
+		ban_user $GLUSER 0 $8 ${10} $7 $0
+		SHOULDKILL=1
+		kill $4 && KILLED=1 && rm /tmp/du-ks/$4
+	}
 	if [ $KILLED -eq 1 ]; then 
 		FORCEKILL=0		
 		i=0
-		while [ -n "$(ps -p $4 -o comm=)" ] && [ $i -lt 4 ]; do
-			i=$(expr $i + 1)
+		while [ -n "`ps -p $4 -o comm=`" ] && [ $i -lt 4 ]; do
+			i=`expr $i + 1`
 			sleep 1
 		done
 
-		[ -n "$(ps -p $4 -o comm=)" ] && O="$O[$(date "+%T %D")] WARNING: process still running after $i seconds, killing by force\n" &&
+		[ -n "`ps -p $4 -o comm=`" ] && { 
+				echo "[`date "+%T %D"`] WARNING: process still running after $i seconds, killing by force" >> $LOG
 				kill -9 $4 && FORCEKILL=1
+		}
 
-    	g_FILE=$(echo $6 | cut -f 2- -d " ")
+    	g_FILE=`echo $6 | cut -f 2- -d " "`
     	[ -n "$g_FILE" ]  && $7 -e dupefile --match "$g_FILE" --loglevel=6 -vvv
     fi    	
 elif [ $SLOW -eq 1 ]; then
-	touch /tmp/du-ks/$4
-elif [ $SLOW -eq 0 ] && [ -f /tmp/du-ks/$4 ]; then
- 	rm /tmp/du-ks/$4
+	touch "/tmp/du-ks/$4"
+elif [ $SLOW -eq 0 ] && [ -f "/tmp/du-ks/$4" ]; then
+ 	rm "/tmp/du-ks/$4"
 fi
 
-[ -n "$O" ] && echo -en $O && [ -n "$LOG" ] && echo -en $O >> $LOG
-
-[ $SHOULDKILL -eq 1 ] && [ $KILLED -eq 0 ] && echo "[$(date "+%T %D")] ERROR: Sending SIGTERM to PID $4 ($GLUSER) failed!" && 
-	[ -f /tmp/du-ks/$4 ] && rm /tmp/du-ks/$4
+[ $SHOULDKILL -eq 1 ] && [ $KILLED -eq 0 ] && echo "[`date "+%T %D"`] ERROR: Sending SIGTERM to PID $4 ($GLUSER) failed!" >> $LOG && 
+	[ -f "/tmp/du-ks/$4" ] && rm "/tmp/du-ks/$4"
 
 exit 1
